+++ common-s3/pom.xml
@@ -11,7 +11,7 @@
     </parent>
 
     <groupId>com.kw</groupId>
-    <artifactId>kotlin.spring-boot--common-s3</artifactId>
+    <artifactId>kotlin-spring-boot--starter-common-s3</artifactId>
     <version>0.0.3-SNAPSHOT</version>
     <name>common-s3</name>
     <description>Starter S3 Library</description>
@@ -30,7 +30,7 @@
         <!-- COMMON LIBRARY -->
         <dependency>
             <groupId>com.kw</groupId>
-            <artifactId>kotlin.spring-boot--common</artifactId>
+            <artifactId>kotlin-spring-boot--starter-common</artifactId>
             <version>${common-starter.version}</version>
             <scope>provided</scope>
         </dependency>
