+++ common/pom.xml
@@ -13,7 +13,7 @@
 	</parent>
 
 	<groupId>com.kw</groupId>
-	<artifactId>kotlin.spring-boot--common</artifactId>
+	<artifactId>kotlin-spring-boot--starter-common</artifactId>
 	<version>3x-SNAPSHOT</version>
 	<name>common</name>
 	<description>Common library for kotlin(+Java) spring boot</description>
