--- common/src/main/kotlin/com/kw/common/starter/constant/ApiOutputStatus.kt
@@ -1,7 +1,0 @@
-package com.kw.common.starter.constant
-
-interface ApiOutputStatus {
-    val code: String
-    val message: String
-    val description: String
-}
