--- common/src/main/kotlin/com/kw/common/starter/decorator/MdcTaskDecorator.kt
@@ -1,20 +1,0 @@
-package com.kw.common.starter.decorator
-
-import org.slf4j.MDC
-import org.springframework.core.task.TaskDecorator
-
-class MdcTaskDecorator : TaskDecorator {
-    override fun decorate(runnable: Runnable): Runnable {
-        val contextMap = MDC.getCopyOfContextMap()
-        return Runnable {
-            try {
-                if (!contextMap.isNullOrEmpty()) {
-                    MDC.setContextMap(contextMap)
-                }
-                runnable.run()
-            } finally {
-                MDC.clear()
-            }
-        }
-    }
-}
