--- common/src/main/kotlin/com/kw/common/starter/exception/AppException.kt
@@ -1,11 +1,0 @@
-package com.kw.common.starter.exception
-
-import com.kw.common.starter.constant.ApiOutputStatus
-import org.springframework.http.HttpStatus
-
-class AppException(
-    val httpStatus: HttpStatus,
-    val apiOutputStatus: ApiOutputStatus,
-    val description: String? = null,
-    override val cause: Throwable? = null,
-) : RuntimeException(cause)
