--- common/src/main/kotlin/com/kw/common/starter/http/filter/HttpRequestFilter.kt
@@ -1,58 +1,0 @@
-package com.kw.common.starter.http.filter
-
-import com.kw.common.starter.http.constant.HeaderFields
-import com.kw.common.starter.http.wrapper.HttpRequestWrapper
-import jakarta.servlet.FilterChain
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.stereotype.Component
-import org.springframework.web.filter.OncePerRequestFilter
-import java.util.UUID
-
-@Component
-class HttpRequestFilter : OncePerRequestFilter() {
-    @Value("\${starter.application-code:APP}")
-    private val applicationCode: String = "APP"
-
-    private val exclusionPaths = listOf("/health", "/info", "/metrics")
-
-    override fun doFilterInternal(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        filterChain: FilterChain,
-    ) {
-        val isExclusionPath = exclusionPaths.any { request.requestURI.endsWith(it) }
-        if (isExclusionPath) {
-            filterChain.doFilter(request, response)
-            return
-        }
-
-        val requestHeaderWrapper = HttpRequestWrapper(request)
-
-        val xCorrelationId = HeaderFields.X_CORRELATION_ID
-
-        HeaderFields.TRACES.forEach { header ->
-            request.getHeader(header).let {
-                if (header == xCorrelationId && it.isNullOrBlank()) initCorrelationId() else it
-            }?.also {
-                requestHeaderWrapper.putHeader(header, it)
-                response.setHeader(header, requestHeaderWrapper.getHeader(header))
-            }
-        }
-
-        filterChain.doFilter(requestHeaderWrapper, response)
-    }
-
-    fun initCorrelationId(): String {
-        val delimiter = "-"
-
-        val randomId =
-            UUID.randomUUID().toString()
-                .split(delimiter)
-                .takeLast(2)
-                .joinToString(separator = delimiter)
-
-        return "$applicationCode-$randomId"
-    }
-}
