--- common/src/main/kotlin/com/kw/common/starter/log/constant/LogbackFields.kt
@@ -1,8 +1,0 @@
-package com.kw.common.starter.log.constant
-
-object LogbackFields {
-    const val CORRELATION_ID = "Correlation-ID"
-    const val TRACE_ID = "X-B3-TraceId"
-    const val SPAN_ID = "X-B3-SpanId"
-    const val SPAN_EXPORT = "X-Span-Export"
-}
