+++ common/src/main/kotlin/com/kw/starter/common/config/properties/ThreadPoolProperty.kt
@@ -1,0 +1,13 @@
+package com.kw.starter.common.config.properties
+
+data class ThreadPoolProperty(
+    val threadNamePrefix: String = "Thd-",
+    val keepAliveTimeSec: Int = 0,
+    val capacity: Int = Int.MAX_VALUE,
+    val pooling: Pooling = Pooling(),
+) {
+    data class Pooling(
+        val core: Int = 1,
+        val max: Int = Int.MAX_VALUE,
+    )
+}
