+++ common/src/main/kotlin/com/kw/starter/common/constant/ApiOutputStatus.kt
@@ -1,0 +1,7 @@
+package com.kw.starter.common.constant
+
+interface ApiOutputStatus {
+    val code: String
+    val message: String
+    val description: String
+}
