+++ common/src/main/kotlin/com/kw/starter/common/exception/AppException.kt
@@ -1,0 +1,11 @@
+package com.kw.starter.common.exception
+
+import com.kw.common.starter.constant.ApiOutputStatus
+import org.springframework.http.HttpStatus
+
+class AppException(
+    val httpStatus: HttpStatus,
+    val apiOutputStatus: ApiOutputStatus,
+    val description: String? = null,
+    override val cause: Throwable? = null,
+) : RuntimeException(cause)
