+++ common/src/main/kotlin/com/kw/starter/common/log/constant/LogbackFields.kt
@@ -1,0 +1,8 @@
+package com.kw.starter.common.log.constant
+
+object LogbackFields {
+    const val CORRELATION_ID = "Correlation-ID"
+    const val TRACE_ID = "X-B3-TraceId"
+    const val SPAN_ID = "X-B3-SpanId"
+    const val SPAN_EXPORT = "X-Span-Export"
+}
