+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.kw</groupId>
-    <artifactId>kotlin.spring-boot--common-libraries</artifactId>
+    <artifactId>kotlin-spring-boot--starter-libraries</artifactId>
     <version>1.0.0-SNAPSHOT</version>
     <packaging>pom</packaging>
